+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Rework changelog
+- Make tests more async
 
 ## 1.0.3 - 2024-01-01